users working and Menu

+++ resources/views/books/show.blade.php
@@ -22,8 +22,6 @@
                                 </p>
                                 <p>{{$book->abstract}}</p>
                                 <hr>
-
-                                @can('admin')
                                      @cannot('guest')
                                         <h4>Documento:</h4>
                                         <div>
@@ -31,6 +29,8 @@
                                             <a  href="{{asset('storage/'.$book->pdfbook)}}" target="_blank"> {{$book->title.'.pdf'}} </a>
                                         </div>
                                     @endcannot
+                                
+                                   @can('admin')
                                     <br>
                                         {!!Form::open(['method'=>'DELETE','route' =>['books.delete',$book],'style'=>'display:inline','onSubmit'=>"return confirm('Are you sure you want to submit this form?');"])!!}
                                         {!!Form::submit('Eliminar_libro',['class'=> 'btn btn-danger']) !!}
